Added Photoshop Files for the Base of the Trees
I created the base of three different trees in Photoshop. I will add branches in my next commit. As a note, for the trees - T1 is middle tree, T2 is left tree, T3 is right tree in my original photo. Feel free to reuse trees as needed to make a larger forest! :)